changed array type of Version from int to byte for more efficiency

+++ Training_Website/IVersionable.cs
@@ -3,6 +3,6 @@
     public interface IVersionable
     {
 
-        void UpdateVersion(int[] version);
+        void UpdateVersion(byte[] version);
     }
 }
